Update of TL sample merging

+++ NightToday/losses.py
@@ -1162,7 +1162,7 @@ def TrafLighLumiLoss_TN(N, T, TN, rec_T, real_D, fake_D, fake_T, mask, contour, 
             luminosity_loss = torch.relu(
                 2. - (color_fake_D * HL_region[b:b + 1] + 2 - 2 * HL_region[b:b + 1])).sum() / (
                                       HL_region[b:b + 1].sum() + 1e-6) + \
-                              torch.relu(-(TN[b:b + 1] * (mask_lighted_area[b:b + 1] - HL_region[b:b + 1])).sum() / (
+                              torch.relu(-(TN[b:b + 1] * (mask_lighted_area[b:b + 1].float() - HL_region[b:b + 1].float())).sum() / (
                                       (mask_lighted_area[b:b + 1] - HL_region[
                                                                     b:b + 1]).sum() + 1e-6) + mean_T_light_region)
             color_dist = ImageTensor(fake_D[b:b + 1] * 0.5 + 0.5).color_distance(
